RAG_local/contextual_compression/requirements.txt
pinecone-client==2.2.4
openai==1.3.0
python-dotenv==1.0.0
tiktoken==0.5.1